chore: remove old nom-based parser code
