+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "montyc",
     "montyc_parser",
-    "montyc_parser2",
     "montyc_core",
     "montyc_driver",
     "montyc_hlirt",
