--- montyc_parser2/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "montyc_parser2"
-version = "0.0.0"
-edition = "2021"
-
-[dependencies]
-montyc_lexer = { path = "../montyc_lexer" }
-montyc_ast = { path = "../montyc_ast" }
-# montyc_query = { path = "../montyc_query" }
-montyc_core = { path = "../montyc_core" }
-
-ariadne = "0.1.5"
-chumsky = "0.8.0"
-ahash = "0.8.0"
-tracing = "0.1.37"
-
-[dev-dependencies]
-serde = "1.0.147"
-serde_json = "1.0.87"
